MTDSA-11940 New package structure - moved connectors to a separate downstream package
