+++ app/api/downstream/common/connectors/connectors.scala
@@ -14,21 +14,14 @@
  * limitations under the License.
  */
 
-package api.connectors
+package api.downstream.common
 
-import config.AppConfig
-import uk.gov.hmrc.http.{HeaderCarrier, HttpClient}
+import api.models.ResponseWrapper
+import api.models.errors.{DownstreamError, MtdError}
 
-import javax.inject.{Inject, Singleton}
-import scala.concurrent.{ExecutionContext, Future}
+package object connectors {
 
-@Singleton
-class MtdIdLookupConnector @Inject() (http: HttpClient, appConfig: AppConfig) {
+  type MtdIdLookupOutcome = Either[MtdError, String]
 
-  def getMtdId(nino: String)(implicit hc: HeaderCarrier, ec: ExecutionContext): Future[MtdIdLookupOutcome] = {
-    import api.connectors.httpparsers.MtdIdLookupHttpParser.mtdIdLookupHttpReads
-
-    http.GET[MtdIdLookupOutcome](s"${appConfig.mtdIdBaseUrl}/mtd-identifier-lookup/nino/$nino")
-  }
-
+  type DownstreamOutcome[A] = Either[ResponseWrapper[DownstreamError], ResponseWrapper[A]]
 }
