+++ test/api/downstream/common/httpparsers/MtdIdLookupHttpParserSpec.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package api.connectors.httpparsers
+package api.downstream.common.httpparsers
 
-import api.connectors.MtdIdLookupOutcome
-import api.connectors.httpparsers.MtdIdLookupHttpParser.mtdIdLookupHttpReads
+import api.downstream.common.connectors.MtdIdLookupOutcome
+import api.downstream.common.httpparsers.MtdIdLookupHttpParser.mtdIdLookupHttpReads
 import api.models.errors.{InvalidBearerTokenError, NinoFormatError, StandardDownstreamError}
 import play.api.libs.json.Writes.StringWrites
 import play.api.libs.json.{JsObject, Json}
