+++ test/api/services/MtdIdLookupServiceSpec.scala
@@ -16,7 +16,7 @@
 
 package api.services
 
-import api.connectors.MockMtdIdLookupConnector
+import api.downstream.common.connectors.MockMtdIdLookupConnector
 import api.models.errors.{NinoFormatError, StandardDownstreamError, UnauthorisedError}
 
 import scala.concurrent.Future
